separate end points for queries
