+++ app/client/src/ce/pages/Editor/IDE/EditorPane/Query/hooks.tsx
@@ -15,7 +15,6 @@ import type { ActionOperation } from "components/editorComponents/GlobalSearch/u
 import { SEARCH_ITEM_TYPES } from "components/editorComponents/GlobalSearch/utils";
 import { createMessage, EDITOR_PANE_TEXTS } from "@appsmith/constants/messages";
 import { getQueryUrl } from "@appsmith/pages/Editor/IDE/EditorPane/Query/utils";
-import { getIDEViewMode, getIsSideBySideEnabled } from "selectors/ideSelectors";
 import {
   ADD_PATH,
   API_EDITOR_ID_PATH,
@@ -26,22 +25,18 @@ import {
 import { SAAS_EDITOR_API_ID_PATH } from "pages/Editor/SaaSEditor/constants";
 import ApiEditor from "pages/Editor/APIEditor";
 import type { UseRoutes } from "@appsmith/entities/IDE/constants";
-import { EditorViewMode } from "@appsmith/entities/IDE/constants";
 import QueryEditor from "pages/Editor/QueryEditor";
 import AddQuery from "pages/Editor/IDE/EditorPane/Query/Add";
-import ListQuery from "pages/Editor/IDE/EditorPane/Query/List";
 import type { AppState } from "@appsmith/reducers";
 import keyBy from "lodash/keyBy";
 import { getPluginEntityIcon } from "pages/Editor/Explorer/ExplorerIcons";
 import { Tag, type ListItemProps } from "design-system";
-import { useCurrentEditorState } from "pages/Editor/IDE/hooks";
 import { createAddClassName } from "pages/Editor/IDE/EditorPane/utils";
 import { QueriesBlankState } from "pages/Editor/QueryEditor/QueriesBlankState";
 
 export const useQueryAdd = () => {
   const location = useLocation();
   const currentEntityInfo = identifyEntityFromPath(location.pathname);
-  const { segmentMode } = useCurrentEditorState();
 
   const openAddQuery = useCallback(() => {
     if (currentEntityInfo.entity === FocusEntity.QUERY_ADD) {
@@ -50,13 +45,13 @@ export const useQueryAdd = () => {
     let url = "";
     url = getQueryUrl(currentEntityInfo);
     history.push(url);
-  }, [currentEntityInfo, segmentMode, location]);
+  }, [currentEntityInfo]);
 
   const closeAddQuery = useCallback(() => {
     let url = "";
     url = getQueryUrl(currentEntityInfo, false);
     history.push(url);
-  }, [currentEntityInfo, segmentMode, location]);
+  }, [currentEntityInfo]);
 
   return { openAddQuery, closeAddQuery };
 };
@@ -107,58 +102,45 @@ export const useGroupedAddQueryOperations = (): GroupedAddOperations => {
   return groups;
 };
 
-export const useQuerySegmentRoutes = (path: string): UseRoutes => {
-  const isSideBySideEnabled = useSelector(getIsSideBySideEnabled);
-  const editorMode = useSelector(getIDEViewMode);
-
-  if (isSideBySideEnabled && editorMode === EditorViewMode.SplitScreen) {
-    return [
-      {
-        key: "ApiEditor",
-        component: ApiEditor,
-        exact: true,
-        path: [
-          BUILDER_PATH + API_EDITOR_ID_PATH,
-          BUILDER_CUSTOM_PATH + API_EDITOR_ID_PATH,
-          BUILDER_PATH_DEPRECATED + API_EDITOR_ID_PATH,
-        ],
-      },
-      {
-        key: "AddQuery",
-        exact: true,
-        component: AddQuery,
-        path: [`${path}${ADD_PATH}`, `${path}/:queryId${ADD_PATH}`],
-      },
-      {
-        key: "SAASEditor",
-        component: QueryEditor,
-        exact: true,
-        path: [
-          BUILDER_PATH + SAAS_EDITOR_API_ID_PATH,
-          BUILDER_CUSTOM_PATH + SAAS_EDITOR_API_ID_PATH,
-          BUILDER_PATH_DEPRECATED + SAAS_EDITOR_API_ID_PATH,
-        ],
-      },
-      {
-        key: "QueryEditor",
-        component: QueryEditor,
-        exact: true,
-        path: [path + "/:queryId"],
-      },
-      {
-        key: "QueryEmpty",
-        component: QueriesBlankState,
-        exact: true,
-        path: [path],
-      },
-    ];
-  }
+export const useQueryEditorRoutes = (path: string): UseRoutes => {
   return [
     {
-      key: "ListQuery",
-      exact: false,
-      component: ListQuery,
-      path: [path, `${path}${ADD_PATH}`, `${path}/:queryId${ADD_PATH}`],
+      key: "ApiEditor",
+      component: ApiEditor,
+      exact: true,
+      path: [
+        BUILDER_PATH + API_EDITOR_ID_PATH,
+        BUILDER_CUSTOM_PATH + API_EDITOR_ID_PATH,
+        BUILDER_PATH_DEPRECATED + API_EDITOR_ID_PATH,
+      ],
+    },
+    {
+      key: "AddQuery",
+      exact: true,
+      component: AddQuery,
+      path: [`${path}${ADD_PATH}`, `${path}/:queryId${ADD_PATH}`],
+    },
+    {
+      key: "SAASEditor",
+      component: QueryEditor,
+      exact: true,
+      path: [
+        BUILDER_PATH + SAAS_EDITOR_API_ID_PATH,
+        BUILDER_CUSTOM_PATH + SAAS_EDITOR_API_ID_PATH,
+        BUILDER_PATH_DEPRECATED + SAAS_EDITOR_API_ID_PATH,
+      ],
+    },
+    {
+      key: "QueryEditor",
+      component: QueryEditor,
+      exact: true,
+      path: [path + "/:queryId"],
+    },
+    {
+      key: "QueryEmpty",
+      component: QueriesBlankState,
+      exact: true,
+      path: [path],
     },
   ];
 };
