+++ app/client/src/pages/Editor/IDE/EditorPane/Explorer.tsx
@@ -2,8 +2,12 @@ import React from "react";
 import { Flex } from "design-system";
 import { Switch, useRouteMatch } from "react-router";
 import { SentryRoute } from "@appsmith/AppRouter";
-import { jsSegmentRoutes } from "@appsmith/pages/Editor/IDE/EditorPane/constants";
+import {
+  jsSegmentRoutes,
+  querySegmentRoutes,
+} from "@appsmith/pages/Editor/IDE/EditorPane/constants";
 import { JSExplorer } from "./JS";
+import { QueryExplorer } from "./Query";
 
 const EditorPaneExplorer = () => {
   const { path } = useRouteMatch();
@@ -18,6 +22,10 @@ const EditorPaneExplorer = () => {
         <SentryRoute
           component={JSExplorer}
           path={jsSegmentRoutes.map((route) => `${path}${route}`)}
+        />
+        <SentryRoute
+          component={QueryExplorer}
+          path={querySegmentRoutes.map((route) => `${path}${route}`)}
         />
       </Switch>
     </Flex>
